http: Move mime-types to http namespace
When making request client will need the ability to configure it
similarly to how server does with replies.
